requirements.txt
APScheduler==3.10.1
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.5.8
Django==5.2.6
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
idna==3.10
marshmallow==3.26.1
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
packaging==25.0
phonepe-sdk==2.1.5
pillow==11.3.0
pycryptodome==3.20.0
PyMySQL==1.1.2
pytz==2025.2
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.12
